exchanged the ExpresionValue wrapper in favor of the new IValueInterpreter to avoid unnecessary conversions; changed the INT TokenType as well as the IntExpression to LONG and a LongExpression to avoid having to differentiate between two more types (int and long); adapted the Interpreter and re-ran tests successfully

+++ src/main/java/me/blvckbytes/gpeee/tokenizer/TokenType.java
@@ -43,7 +43,7 @@ public enum TokenType {
   }),
 
   // -?[0-9]+
-  INT(TokenCategory.VALUE, null, tokenizer -> {
+  LONG(TokenCategory.VALUE, null, tokenizer -> {
     StringBuilder result = new StringBuilder();
 
     if (collectDigits(tokenizer, result, false) != CollectorResult.READ_OKAY)
